Bug fix for a bug reported by Nuno Lopes in PUSHREF.
Basically if the argument to PUSHREF is evaluated always before the computation of the lvalue in the macro, all is fine.  But this order of evaluation seems to be undefined by the standard, and apparently differs depending on the compiler.

The fix is turning the macro into a static inline function.